requirements.txt
absl-py==0.10.0
adafruit-circuitpython-busdevice==4.3.1
adafruit-circuitpython-dotstar==2.0.0
adafruit-circuitpython-fxas21002c==2.1.5
adafruit-circuitpython-fxos8700==2.1.4
fake_rpi==0.7.1
gps==3.19
grpcio==1.18.0
LabJackPython==2.0.0
mlx9064x-driver==1.3.0
mock==4.0.2
protobuf==3.12.2
psycopg2==2.7.7
pygatt==4.0.5
pylint==2.6.0
python-geohash==0.8.5
pytype==2020.8.28
PyYAML==5.4
RPi.GPIO==0.7.0
seaborn==0.9.0
sdnotify==0.3.2
testing.postgresql==1.3.0

# scipy deps
Cython==0.29.21  # For pandas
matplotlib==3.0.2
ipython==5.8.0
pandas==0.23.3

# sckkit-learn deps first.
numpy==1.16.2
# Finally scipy & scikit.
scipy==1.4.1
scikit-learn==0.20.2
